+++ analytical/tests/__init__.py
@@ -7,6 +7,7 @@ from analytical.tests.test_tag_chartbeat import *
 from analytical.tests.test_tag_clicky import *
 from analytical.tests.test_tag_crazy_egg import *
 from analytical.tests.test_tag_google_analytics import *
+from analytical.tests.test_tag_gosquared import *
 from analytical.tests.test_tag_hubspot import *
 from analytical.tests.test_tag_kiss_insights import *
 from analytical.tests.test_tag_kiss_metrics import *